backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
google-generativeai==0.3.2
elevenlabs==0.2.26
cloudinary==1.38.0
python-dotenv==1.0.0
httpx==0.25.2
edge-tts